requirements.txt
langgraph==0.2.74
langchain==0.3.23
langchain-openai==0.3.12
langchain-anthropic==0.3.3
langchain-google-genai==2.0.0
langchain-community==0.3.19
pydantic==2.10.6
requests==2.32.3
python-dotenv==1.0.1
tavily-python==0.5.1
spacy==3.8.5
nltk==3.9.1
beautifulsoup4==4.12.3
newsapi-python==0.2.7
fastapi==0.115.11
uvicorn==0.34.0
google-generativeai==0.7.2
duckduckgo-search==8.0.0
wikidata==0.7.0 